Move controller_app to controller/webapp
