+++ controller/webapp/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for controller_app project.
+Django settings for controller.webapp project.
 
 Generated by 'django-admin startproject' using Django 5.1.5.
 
@@ -63,7 +63,7 @@ DEBUG = get_env_var("DJANGO_DEBUG") == "True"
 
 INSTALLED_APPS = [
     "django.contrib.staticfiles",
-    "controller_app",
+    "controller.webapp",
 ]
 
 MIDDLEWARE = [
@@ -73,7 +73,7 @@ MIDDLEWARE = [
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
 ]
 
-ROOT_URLCONF = "controller_app.urls"
+ROOT_URLCONF = "controller.webapp.urls"
 
 TEMPLATES = [
     {
@@ -91,7 +91,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = "controller_app.wsgi.application"
+WSGI_APPLICATION = "controller.webapp.wsgi.application"
 
 
 # Database
