+++ controller/webapp/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for controller_app project.
+WSGI config for controller.webapp project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -12,6 +12,6 @@ import os
 from django.core.wsgi import get_wsgi_application
 
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "controller_app.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "controller.webapp.settings")
 
 application = get_wsgi_application()
