+++ docker/docker-compose.yaml
@@ -118,7 +118,7 @@ services:
   test-controller-web:
     extends:
         service: test-controller
-    command: gunicorn --config gunicorn.conf.py controller_app.wsgi
+    command: gunicorn --config gunicorn.conf.py controller.webapp.wsgi
 
   test-controller-service:
     extends:
