+++ manage.py
@@ -10,7 +10,7 @@ from jobrunner.lib.log_utils import configure_logging
 
 def main():
     """Run administrative tasks."""
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "controller_app.settings")
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "controller.webapp.settings")
     tracing.setup_default_tracing()
     configure_logging()
